fixes custom emoji not appearing in autocomplete
