+++ app/javascript/mastodon/features/emoji/emoji_picker.tsx
@@ -1,39 +1,19 @@
 import type { FC } from 'react';
-import { useEffect, useState } from 'react';
 
-import type {
-  CategoryName,
-  CustomEmoji,
-  EmojiProps,
-  PickerProps,
-} from 'emoji-mart';
+import type { EmojiProps, PickerProps } from 'emoji-mart';
 import EmojiRaw from 'emoji-mart/dist-es/components/emoji/nimble-emoji';
 import PickerRaw from 'emoji-mart/dist-es/components/picker/nimble-picker';
 
-import { autoPlayGif } from '@/mastodon/initial_state';
-import { assetHost } from 'mastodon/utils/config';
+import { assetHost } from '@/mastodon/utils/config';
 
 import { EMOJI_MODE_NATIVE } from './constants';
 import EmojiData from './emoji_data.json';
 import { useEmojiAppState } from './mode';
-import { emojiLogger } from './utils';
+import { usePickerEmojis } from './picker';
 
 const backgroundImageFnDefault = () => `${assetHost}/emoji/sheet_16_0.png`;
 
-let customEmojis: CustomEmoji[] | null = null;
-let customCategories = [
-  'recent',
-  'people',
-  'nature',
-  'foods',
-  'activity',
-  'places',
-  'objects',
-  'symbols',
-  'flags',
-] as CategoryName[];
-
-const log = emojiLogger('picker');
+export { fetchCustomEmojiData as loadCustomEmojiData } from './picker';
 
 export const Picker: FC<PickerProps> = ({
   set = 'twitter',
@@ -44,24 +24,16 @@ export const Picker: FC<PickerProps> = ({
   ...props
 }) => {
   const { mode } = useEmojiAppState();
-  const [isLoaded, setLoaded] = useState(customEmojis !== null);
+  const { customCategories, customEmojis } = usePickerEmojis();
 
-  useEffect(() => {
-    if (customEmojis === null) {
-      void loadCustomEmojiData().then(() => {
-        setLoaded(true);
-      });
-    }
-  }, []);
-
-  if (!isLoaded) {
+  if (!customEmojis) {
     return null;
   }
 
   return (
     <PickerRaw
       data={EmojiData}
-      custom={customEmojis ?? []}
+      custom={customEmojis}
       include={customCategories}
       set={set}
       sheetSize={sheetSize}
@@ -73,46 +45,6 @@ export const Picker: FC<PickerProps> = ({
     />
   );
 };
-
-export async function loadCustomEmojiData() {
-  const { loadAllCustomEmoji } = await import('./database');
-  const emojisRaw = await loadAllCustomEmoji();
-  if (emojisRaw.length === 0) {
-    return;
-  }
-
-  const categories = new Set(['custom']);
-  const emojis = [];
-  for (const emoji of emojisRaw) {
-    const name = emoji.shortcode.replaceAll(':', '');
-    emojis.push({
-      name,
-      id: name,
-      custom: true,
-      short_names: [name],
-      imageUrl: autoPlayGif ? emoji.url : emoji.static_url,
-      customCategory: emoji.category,
-    });
-
-    if (emoji.category) {
-      categories.add(`custom-${emoji.category}`);
-    }
-  }
-
-  customEmojis = emojis.toSorted((a, b) => {
-    return a.name.toLowerCase().localeCompare(b.name.toLowerCase());
-  });
-  customCategories = customCategories.toSpliced(
-    1,
-    0,
-    ...(Array.from(categories).toSorted() as CategoryName[]),
-  );
-  log(
-    'loaded %d custom emojis in %d categories',
-    customEmojis.length,
-    categories.size,
-  );
-}
 
 export const Emoji: FC<EmojiProps> = ({
   set = 'twitter',
